Add files via upload
This project demonstrates a basic image steganography technique using Least Significant Bit (LSB) manipulation in Python. The goal is to hide a secret text message inside an image without visibly altering the image, ensuring secure and covert communication. The project runs on Google Colab and uses the Pillow library for image processing.

The user uploads a .png image and enters a message to be hidden. The program encodes this message by modifying the least significant bits of the image's RGB pixel values. Later, the same or another user can decode the message using the provided function.